test(e2e): wait logic moved from is-faucet-ready.ts script to the faucet start method
