+++ packages/e2e/package.json
@@ -44,8 +44,7 @@
     "test:build:verify": "yarn test:web-extension:build",
     "test:debug": "DEBUG=true yarn test",
     "generate-mnemonics": "ts-node src/util/mnemonic.ts",
-    "wait-for-network": "ts-node src/util/is-local-network-ready.ts",
-    "wait-for-faucet": "ts-node src/util/is-faucet-ready.ts"
+    "wait-for-network": "ts-node src/util/is-local-network-ready.ts"
   },
   "repository": "https://github.com/input-output-hk/cardano-js-sdk",
   "contributors": [
